segregated resolvers
