+++ server/graphql/resolvers/Author/Mutation.js
@@ -1,0 +1,13 @@
+const authorService = require("../../../services/AuthorService");
+
+module.exports = {
+    createAuthor: (_, { payload }) => {
+        return authorService.createAuthor(payload);
+    },
+    updateAuthor: (_, { id, payload }) => {
+        return authorService.updateAuthor(id, payload);
+    },
+    deleteAuthor: (_, { id }) => {
+        return authorService.deleteAuthor(id);
+    },
+}
