+++ server/graphql/resolvers/Author/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    Mutation: require('./Mutation'),
+    Query: require('./Query'),
+    Resolver: require('./Resolver')
+}
