+++ server/graphql/resolvers/AuthorReview/Mutation.js
@@ -1,0 +1,7 @@
+const authorReviewService = require("../../../services/AuthorReviewService");
+
+module.exports = {
+    createAuthorReview: (_, { authorId, payload }) => {
+        return authorReviewService.createAuthorReview(authorId, payload);
+    },
+}
