+++ server/graphql/resolvers/AuthorReview/Resolver.js
@@ -1,0 +1,7 @@
+module.exports = {
+    AuthorReview: {
+        author: (parent, _, ctx) => {
+            return ctx.authorLoader.load(parent.author_id);
+        }
+    },
+}
