+++ server/graphql/resolvers/BookReview/Mutation.js
@@ -1,0 +1,7 @@
+const bookReviewService = require("../../../services/BookReviewService");
+
+module.exports = {
+    createBookReview: (_, { bookId, payload }) => {
+        return bookReviewService.createBookReview(bookId, payload);
+    }
+}
