+++ server/graphql/resolvers/BookReview/Resolver.js
@@ -1,0 +1,7 @@
+module.exports = {
+    BookReview: {
+        book: (parent, _, ctx) => {
+            return ctx.bookLoader.load(parent.book_id);
+        },
+    }
+}
