+++ server/graphql/resolvers/BookReview/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    Mutation: require('./Mutation'),
+    Resolver: require('./Resolver'),
+}
